% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d.morph_no_ringnr}
\alias{d.morph_no_ringnr}
\title{d.morph data}
\format{
A data frame with 4625 rows and 24 variables:
\describe{
  \item{ID}{int: ID of the individual}
  \item{indnr}{int: Individual number}
  \item{sex}{num: Sex of the individual}
  \item{island_current}{Factor: Current island}
  \item{adultSNPisland}{int: Adult SNP island}
  \item{hatchisland}{int: Hatch island}
  \item{hatchyear}{Factor: Hatch year}
  \item{year}{Factor: Year}
  \item{month}{num: Month}
  \item{age}{num: Age}
  \item{tarsus}{num: Tarsus length}
  \item{wing}{num: Wing length}
  \item{mass}{num: Mass}
  \item{island}{int: Island}
  \item{FGRM}{num: FGRM value}
  \item{id}{int: ID}
  \item{billD}{num: Bill depth}
  \item{inner}{num: Inner measurement}
  \item{outer}{num: Outer measurement}
  \item{other}{num: Other measurement}
  \item{IDC}{int: IDC value}
  \item{IDC2}{int: IDC2 value}
  \item{IDC3}{int: IDC3 value}
  \item{IDC4}{int: IDC4 value}
}
}
\source{
Data from sparrows on the Helgeland coast, Norway
}
\usage{
d.morph_no_ringnr
}
\description{
A dataset containing morphological measurements of a house sparrow population on the coast of Helgeland, Norway.
}
\keyword{datasets}
