% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{extract_effects}
\alias{extract_effects}
\title{Extract Fixed and Random Effects from a Model Formula}
\usage{
extract_effects(formula)
}
\arguments{
\item{formula}{A model formula.}
}
\value{
A list with two elements: `fixed_effects` and `random_effects`,
        each containing the names of the respective effects extracted from the formula.
}
\description{
This function identifies and separates fixed and random effects specified in a model formula.
}
\examples{
formula <- y ~ x + f(group, model = "iid")
effects <- extract_effects(formula)
}
