% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{extract_importances}
\alias{extract_importances}
\title{Extract Importance of Fixed and Random Effects}
\usage{
extract_importances(model, data, dist_factor = NULL)
}
\arguments{
\item{model}{The fitted model object.}

\item{data}{The data used for fitting the model.}

\item{dist_factor}{Distributional variance factor, used to calculate residual variance. If NULL, the family and link will be checked. If the family is binomial with logit link, the dist_factor will be set to pi^2 / 3. If the family is binomial with probit link, the dist_factor will be set to 1. If the family is poisson, the dist_factor will be calculated as described in Nakagawa et. al - A general and simple method for obtaining R2 from generalized linear mixed-effects models (2013) and Nakagawa et. al - The coefficient of determination R2 and intra-class correlation coefficient from generalized linear mixed-effects models revisited and expanded (2017). If the family does not match any of these, the dist_factor will be set to 0, corresponding to Gaussian responses (LMM)}
}
\value{
A list containing normalized importance values for random and fixed effects, marginal and conditional R2 values, and expected importances.
}
\description{
Calculates the relative importance of both fixed and random effects from a fitted INLA model.
}
\examples{
# Assuming `model` is an INLA model object with log, logit or probit link and `data` contains appropriate predictors
importance <- extract_importances(model, data, dist_factor = pi^2 / 3)
}
