% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_posteriors_and_heritability}
\alias{plot_posteriors_and_heritability}
\title{Plot Posterior Distributions of Variance Components and Heritability}
\usage{
plot_posteriors_and_heritability(model, random_effect_name = NULL)
}
\arguments{
\item{model}{An `inla` model object.}

\item{random_effect_name}{Optional name of the random effect for which heritability is to be plotted.}
}
\value{
A list of plot objects: `variance_plot` for variance components and `heritability_plot` for heritability.
}
\description{
Generates plots for the posterior distributions of variance components and, optionally, heritability.
}
\examples{
# Assuming 'result' is an inla model object
plots <- plot_posteriors_and_heritability(result, random_effect_name = "group")
}
