% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_samples}
\alias{plot_samples}
\title{Plot Samples from Posterior Distributions}
\usage{
plot_samples(samples)
}
\arguments{
\item{samples}{A list object containing sampled values and derived quantities from `sample_posterior`.}
}
\value{
A list of ggplot objects for each of the plotted quantities.
}
\description{
Generates plots for the scaled importance of fixed and random effects,
as well as marginal and conditional R2 values from sampled posterior distributions.
}
\examples{
# Assuming 'samples' contains results from `sample_posterior`
plots <- plot_samples(samples)
}
