% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{sample_posterior_count}
\alias{sample_posterior_count}
\title{Sample Posterior Distributions for Non-Gaussian Model Parameters}
\usage{
sample_posterior_count(
  model,
  formula,
  data,
  n_samp = 1000,
  additive_param = NULL,
  repeatability = FALSE
)
}
\arguments{
\item{model}{The fitted model object from INLA.}

\item{formula}{The model formula.}

\item{data}{The data used for fitting the model.}

\item{n_samp}{Number of samples to draw from the posterior distributions.}

\item{additive_param}{Optional; specifies an additive parameter for which to compute repeatability}

\item{repeatability}{Optional; calculates the repeatability (variance of additive_param divided by sum of random variances) according to Stoffel et. al - rptR: repeatability estimation and variance decomposition by generalized linear mixed-effects models (2017).}
}
\value{
A list containing matrices and data frames of sampled values and derived quantities.
}
\description{
Samples from the posterior distributions of model parameters and computes derived quantities for non-Gaussian models.
}
\examples{
# Assuming 'result' is an INLA model object and 'data_binomial' is available
samples <- sample_posterior_count(result, formula, data_binomial, n_samp = 100)
}
